# specs/005-content-personalization/contracts/api.yaml
openapi: 3.1.0
info:
  title: Content Personalization API
  version: 1.0.0
  description: API for user management, onboarding, and content personalization
servers:
  - url: http://localhost:8000/api
    description: Local development server
paths:
  /user/create:
    post:
      summary: Create a new anonymous user
      operationId: createUser
      tags:
        - User
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '409':
          description: User with this ID already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /user/{user_id}:
    get:
      summary: Get user information
      operationId: getUser
      tags:
        - User
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    CreateUserRequest:
      type: object
      required:
        - user_id
      properties:
        user_id:
          type: string
          format: uuid
          description: Client-generated UUID for the user
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        onboarding_completed:
          type: boolean
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error type
        message:
          type: string
          description: Human-readable error message
        detail:
          type: object
          description: Additional error details